day-09/fastapi_app/requirements.txt
fastapi
uvicorn
routers
psutils
router
boto3